[fix]修复Android 10以下机型双指操作播放界面时崩溃的问题；多渠道打包；bugly更换为u-apm

+++ app/src/main/java/com/skyd/imomoe/util/html/source/web/GettingUtil.kt
@@ -19,8 +19,8 @@ class GettingUtil private constructor() {
     private lateinit var mCallback: GettingCallback
     private var mHeader: MutableMap<String, String> = HashMap()
     private var mActivity: SoftReference<Activity>? = null
-    private var mConnTimeOut = 16 * 1000.toLong()
-    private var mReadTimeOut = 20 * 1000.toLong()
+    private var mConnTimeOut = 16 * 1000L
+    private var mReadTimeOut = 20 * 1000L
     private var mFinishedTimeOut: Long = 800
     private lateinit var mClient: GettingWebViewClient
